Move TestAssignable into thema package
